Add componentWillMount method to update state
I got a warning and tests will fail for that
that's why I created the componentWillMount method
here is the warning:
```
Warning: setState(...): Cannot update during an existing state
transition (such as within `render` or another component's constructor).
Render methods should be a pure function of props and state; constructor
side-effects are an anti-pattern, but can be moved to
`componentWillMount`.
```