Allow the use of the proxy for adding layers
If the geonode and the geoserver are on different ports or different
servers (very unlikely) CORS is needed for the new viewer/composer to
work. The different ports is the default setting for a development
environment. Therefore we need the ability to use the geonode proxy to
relay the GetCapabilities and the Layers through it.
The SDK got updated to support this for GetCapabilities and this commit
adds the support for the GeoNode Composer