Add the ability to use any proxy (#82)
* Allow the use of the proxy for adding layers

If the geonode and the geoserver are on different ports or different
servers (very unlikely) CORS is needed for the new viewer/composer to
work. The different ports is the default setting for a development
environment. Therefore we need the ability to use the geonode proxy to
relay the GetCapabilities and the Layers through it.
The SDK got updated to support this for GetCapabilities and this commit
adds the support for the GeoNode Composer

* Use proxy in templates for geonode

Update sdk and add proxy to MapTransform

* Fix test due to new alignment

maybe a good sign to remove the test