Edit map (#116)
Allow for editing maps with the composer. Setting the mapId and the mapConfig will allow the client to sent a PUT request to update the map instead of a POST to create a new map.
I reorganized the Redux actions and reducers to clarify to whom they belong. Add new selectors for easy access to state.
Add integration tests for Redux actions/reducers/selectors
* Move save to geonode endpoint into a service
* add rewire to mock the service library for the action tests
* Refactor redux actions into state folders
* Add new selectors for map
selectors separates the way of getting the state for specific items
just thorugh methods instead of fiddling with the internal way the state
is stored
* add integration tests for reducers, actions and selectors
* Add enzyme and tresdouble for testing components
* Split the SaveDetailModal into two components
split it into two components, one container handling the actions and one
view handling the rendering
makes it easier to test and replace certain components
* Add new service method for transforming the map into a config
* remove old reducers and actions use the new organized reducers and
actions and selector
* set mapId for the composer to start editing